Battery class introduced

+++ Source/Encoder.fs
@@ -3,7 +3,7 @@ open System
 open Trik
 type Encoder(i2cCommandNumber) =
     inherit Internals.PollingSensor<int>()
-    override self.Read() = Helpers.I2C.Receive i2cCommandNumber
+    override self.Read() = int(int16(Helpers.I2C.Receive i2cCommandNumber))
     member self.Reset() = Helpers.I2C.Send i2cCommandNumber 0 2
     interface IDisposable with
         member x.Dispose() = ()
